android_app/android_backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
catboost==1.2.5
certifi==2024.2.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
distro==1.9.0
Flask==3.0.3
fonttools==4.51.0
graphviz==0.20.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
numpy==1.26.4
openai==1.23.2
packaging==24.0
pandas==2.2.2
pillow==10.3.0
plotly==5.21.0
pydantic==2.7.0
pydantic_core==2.18.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
scipy==1.13.0
six==1.16.0
sniffio==1.3.1
tenacity==8.2.3
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
Werkzeug==3.0.2
